Update ESLint configuration and upgrade dependencies
This commit includes updates to the ESLint configuration and its related packages. React plugin has been added and unused packages have been removed from the ESLint configuration. It also includes the introduction of an ESLint rule for unescaped entities in React and updating the corresponding version in the changelog. In addition, the .gitignore file has been updated to include the .idea folder. Lastly, the package and lock files were updated reflecting the updated package versions and dependency tree.